fix(core,parsers,server): apply CTO audit remediations and add real benchmark harness
- Security: fix ValidatedScope sibling prefix bypass using component-based Path::starts_with
- Governance: unblock read-only smart_search in contract repositories (proto-registry)
- Scalability: upgrade RepoId from u8 to u16 (65,535 repositories support)
- Parsers: add 1.5 MB AstGuard budget for schemas/stubs and context-aware lexical nesting depth guard
- Concurrency: add advisory OS file locking (libc::flock) to audit.log hash chain appends
- Tests: expand test suite from 33 to 48 comprehensive unit/integration tests
- Benchmarks: implement reproducible real-world benchmark suite (cargo bench) measuring actual latencies and token savings

+++ crates/mesh-core/src/types.rs
@@ -4,7 +4,7 @@ use std::path::PathBuf;
 /// Re-export CompactString as CompactStr for zero-allocation hot loops per RFC-001 Commandment 1
 pub type CompactStr = compact_str::CompactString;
 
-pub type RepoId = u8;
+pub type RepoId = u16;
 
 pub type SymbolName = CompactStr;
 pub type PathStr = CompactStr;
